Renames SmallGraceCompiler as GraceCompiler.  Removes module methods, and creates module files.  Automatically compiles imported modules if they have not been compiled.
TODO: check that existing compiled modules are up-to-date with their source.

Improves scope-building for types.  Removes type-parameters from set of names provide by a type by adding the `isAvailableForReuse` predicate.

Implements scope-retrival for dialects.

Changes grammar to replace Identifiers in type expressions with requests.  Links scanner to parser in addition to the existing linkage from parser to scanner, so that tokens know which module they are in.  This helps to ensure that syntactic errors are reported on the correct module.