Introduces all annotations in the standard modules.
Renames modules containing classes and traits with the suffix "Bundle", and names the generator method "open".

Since method bodie sand def initializers are now optional in the grammar, the BuildScopes visitor checks that they are present unless the declaration is annotated with a marker method.

Conflicting language-defined annotations, principally confidential and pubic, are now detected.