Fixes "outer" issues.  All tests green.
I decided that I cannot get rid of <selfDot> and <outerDot> as tokens.
We also need <self> to be a token, so that it can have an Id.   This means that patching the token Actions table to give reserved words priority over identifiers must look for  <reserved> as well as for "reserved".