Adds where clasues (untested) and once methods.
Adds tests for the visibility annotations and for the presence of the `once` tag on methods.
No tests for `where` as yet, because I don't know what it should do.  But `where` is a
reserved word, and I wanted it to be treated as such.

`super` is no longer a reserved word.

This commit was pieced together by hand after git created a nightmare.